+++ packages/types/src/constants.ts
@@ -22,3 +22,4 @@ export type NetworkId = "testnet" | "mainnet";
 
 export const ARBITRUM_TESTNET_CHAINID = 421613;
 export const ARBITRUM_MAINNET_CHAINID = 42161;
+export const ARBITRUM_MAINNET_CHAINID_HEX = "0xa4b1";