+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,37 @@
 # examples
 
+## 0.1.34
+
+### Patch Changes
+
+- update
+- update
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/components@0.0.65
+  - @orderly.network/hooks@0.0.66
+  - @orderly.network/core@0.0.64
+
+## 0.1.33
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.64
+  - @orderly.network/hooks@0.0.65
+  - @orderly.network/core@0.0.63
+
+## 0.1.32
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.63
+  - @orderly.network/core@0.0.62
+  - @orderly.network/hooks@0.0.64
+
 ## 0.1.31
 
 ### Patch Changes