+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,24 @@
 # storybook
 
+## 1.0.51
+
+### Patch Changes
+
+- update
+- update
+
+## 1.0.50
+
+### Patch Changes
+
+- update
+
+## 1.0.49
+
+### Patch Changes
+
+- update
+
 ## 1.0.48
 
 ### Patch Changes