+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/cli
 
+## 0.0.9
+
+### Patch Changes
+
+- update
+- update
+
+## 0.0.8
+
+### Patch Changes
+
+- update
+
+## 0.0.7
+
+### Patch Changes
+
+- update
+
 ## 0.0.6
 
 ### Patch Changes