+++ packages/component/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/components
 
+## 0.0.65
+
+### Patch Changes
+
+- update
+- update
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/types@0.0.64
+  - @orderly.network/utils@0.0.52
+  - @orderly.network/net@1.0.61
+
+## 0.0.64
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.63
+  - @orderly.network/utils@0.0.51
+  - @orderly.network/net@1.0.60
+
+## 0.0.63
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.59
+  - @orderly.network/types@0.0.62
+  - @orderly.network/utils@0.0.50
+
 ## 0.0.62
 
 ### Patch Changes