+++ packages/core/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/core
 
+## 0.0.64
+
+### Patch Changes
+
+- update
+- update
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/types@0.0.64
+
+## 0.0.63
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.63
+
+## 0.0.62
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.62
+
 ## 0.0.61
 
 ### Patch Changes