+++ packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,24 @@
 # eslint-config-custom
 
+## 0.0.5
+
+### Patch Changes
+
+- update
+- update
+
+## 0.0.4
+
+### Patch Changes
+
+- update
+
+## 0.0.3
+
+### Patch Changes
+
+- update
+
 ## 0.0.2
 
 ### Patch Changes