+++ packages/futures/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @orderly.network/futures
 
+## 0.0.65
+
+### Patch Changes
+
+- update
+- update
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/types@0.0.64
+  - @orderly.network/utils@0.0.52
+
+## 0.0.64
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.63
+  - @orderly.network/utils@0.0.51
+
+## 0.0.63
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.62
+  - @orderly.network/utils@0.0.50
+
 ## 0.0.62
 
 ### Patch Changes