+++ packages/net/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/net
 
+## 1.0.61
+
+### Patch Changes
+
+- update
+- update
+
+## 1.0.60
+
+### Patch Changes
+
+- update
+
+## 1.0.59
+
+### Patch Changes
+
+- update
+
 ## 1.0.58
 
 ### Patch Changes