+++ packages/types/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/types
 
+## 0.0.64
+
+### Patch Changes
+
+- update
+- update
+
+## 0.0.63
+
+### Patch Changes
+
+- update
+
+## 0.0.62
+
+### Patch Changes
+
+- update
+
 ## 0.0.61
 
 ### Patch Changes