+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/utils
 
+## 0.0.52
+
+### Patch Changes
+
+- update
+- update
+
+## 0.0.51
+
+### Patch Changes
+
+- update
+
+## 0.0.50
+
+### Patch Changes
+
+- update
+
 ## 0.0.49
 
 ### Patch Changes